requirements.txt
anthropic>=0.58.0
openai>=1.0.0
wandb>=0.16.0
python-dotenv>=1.0.0
tqdm>=4.67.0
PyYAML>=6.0.0
GitPython>=3.1.0
huggingface-hub>=0.24.0
hf-transfer>=0.1.5

